aggressively refactoring, pruning old work to push forward from current state
