+++ skyrim-extractor/package.json
@@ -12,10 +12,10 @@
     "dev": "ts-node src/cli.ts --config ./config.json",
     "dev:debug": "ts-node src/cli.ts --config ./config.json --debug",
     "test": "jest",
-    "extract": "node dist/cli.js --config ./config.json",
-    "extract:debug": "node dist/cli.js --config ./config.json --debug",
-    "extract:refactor": "ts-node src/index.refactor.ts --config ./config.json",
-    "extract:refactor:debug": "ts-node src/index.refactor.ts --config ./config.json --debug"
+
+    "extract": "ts-node src/index.ts --config ./config.json",
+    "extract:debug": "ts-node src/index.ts --config ./config.json --debug",
+    "extract:test": "jest"
   },
   "dependencies": {
     "@lorerim/platform-types": "file:../__platform",
