--- skyrim-extractor/src/binary-decoders/grup/grupUtils.ts
@@ -1,204 +1,0 @@
-import { parentPort } from "worker_threads";
-import { Buffer } from "buffer";
-import { ParsedRecord } from "@lorerim/platform-types";
-import { ProcessedRecordType } from "../../constants";
-import { RECORD_HEADER } from "../buffer.constants";
-import {
-  parseRecordHeader,
-  scanSubrecords,
-  extractSubrecordDataAsBase64,
-} from "../recordParser";
-import * as fs from "fs";
-import * as path from "path";
-import { hexDump } from "../bufferParser";
-
-// Debug file for GRUP dumps
-const GRUP_DEBUG_FILE = path.join(process.cwd(), "grups.debug");
-
-/**
- * Send a debug message to the main process
- */
-export function debugLog(message: string) {
-  parentPort?.postMessage({ type: "debug", message });
-}
-
-/**
- * Send an error message to the main process
- */
-export function errorLog(functionName: string, message: string) {
-  parentPort?.postMessage({
-    type: "error",
-    error: `[${functionName}] ${message}`,
-  });
-}
-
-/**
- * Get the name of a GRUP type
- */
-export function getGroupTypeName(type: number): string {
-  const types: Record<number, string> = {
-    0: "Top-Level",
-    1: "World Children",
-    2: "Interior Cell Block",
-    3: "Interior Cell Sub-Block",
-    4: "Exterior Cell Block",
-    5: "Exterior Cell Sub-Block",
-    6: "Cell Children",
-    7: "Topic Children",
-    8: "Cell Persistent Children",
-    9: "Cell Temporary Children",
-  };
-  return types[type] || `Unknown (${type})`;
-}
-
-/**
- * Parse a GRUP header from a buffer
- */
-export function parseGRUPHeader(buffer: Buffer, offset: number) {
-  if (buffer.length < offset + RECORD_HEADER.TOTAL_SIZE) {
-    throw new Error(
-      `Buffer too small for GRUP header: ${buffer.length} < ${
-        offset + RECORD_HEADER.TOTAL_SIZE
-      }`
-    );
-  }
-
-  const header = buffer.slice(offset, offset + RECORD_HEADER.TOTAL_SIZE);
-  const type = header.toString("ascii", 0, 4);
-  if (type !== "GRUP") {
-    throw new Error(`Invalid GRUP header type: ${type}`);
-  }
-
-  return {
-    type: "GRUP",
-    size: header.readUInt32LE(4),
-    groupType: header.readUInt32LE(8),
-    groupTypeStr: getGroupTypeName(header.readUInt32LE(8)),
-    label: header.slice(12, 16),
-    timestamp: header.readUInt32LE(16),
-    versionControl: header.readUInt32LE(20),
-  };
-}
-
-/**
- * Validate GRUP size
- */
-export function validateGRUPSize(header: any, buffer: Buffer, offset: number) {
-  if (header.size < RECORD_HEADER.TOTAL_SIZE) {
-    throw new Error(
-      `GRUP size too small: ${header.size} < ${RECORD_HEADER.TOTAL_SIZE}`
-    );
-  }
-  if (offset + header.size > buffer.length) {
-    throw new Error(
-      `GRUP size exceeds buffer length: ${offset + header.size} > ${
-        buffer.length
-      }`
-    );
-  }
-}
-
-/**
- * Process a GRUP record
- * @param buffer The buffer containing the GRUP record
- * @param offset The offset to start processing from
- * @param pluginName The name of the plugin being processed
- * @param processedTypes Set of record types to process
- * @returns The processed records and the new offset
- */
-export function processGrupRecord(
-  buffer: Buffer,
-  offset: number,
-  pluginName: string,
-  processedTypes: Set<ProcessedRecordType>
-): { records: ParsedRecord[]; newOffset: number } {
-  const records: ParsedRecord[] = [];
-  let currentOffset = offset;
-
-  // Read GRUP header
-  const grupHeader = buffer.slice(
-    currentOffset,
-    currentOffset + RECORD_HEADER.TOTAL_SIZE
-  );
-  const header = parseRecordHeader(grupHeader);
-  currentOffset += RECORD_HEADER.TOTAL_SIZE;
-
-  // Process all records in the GRUP
-  while (currentOffset < offset + header.dataSize) {
-    const recordHeader = parseRecordHeader(
-      buffer.slice(currentOffset, currentOffset + RECORD_HEADER.TOTAL_SIZE)
-    );
-
-    // Skip unsupported record types
-    if (!processedTypes.has(recordHeader.type as ProcessedRecordType)) {
-      currentOffset += RECORD_HEADER.TOTAL_SIZE + recordHeader.dataSize;
-      continue;
-    }
-
-    const data = buffer.subarray(
-      currentOffset + RECORD_HEADER.TOTAL_SIZE,
-      currentOffset + RECORD_HEADER.TOTAL_SIZE + recordHeader.dataSize
-    );
-
-    const subrecords: Record<string, string[]> = {};
-    let subrecordOffset = 0;
-    for (const subrecord of scanSubrecords(data, 0).subrecords) {
-      if (!subrecords[subrecord.header.type]) {
-        subrecords[subrecord.header.type] = [];
-      }
-
-      subrecords[subrecord.header.type].push(
-        extractSubrecordDataAsBase64(
-          data,
-          subrecordOffset,
-          subrecord.header.size
-        )
-      );
-      subrecordOffset += subrecord.header.size;
-    }
-
-    records.push({
-      meta: {
-        type: recordHeader.type,
-        formId: recordHeader.formId.toString(16).toUpperCase().padStart(8, "0"),
-        plugin: pluginName,
-      },
-      data: subrecords,
-      header: buffer
-        .slice(currentOffset, currentOffset + RECORD_HEADER.TOTAL_SIZE)
-        .toString("base64"),
-    });
-
-    currentOffset += RECORD_HEADER.TOTAL_SIZE + recordHeader.dataSize;
-  }
-
-  return {
-    records,
-    newOffset: offset + RECORD_HEADER.TOTAL_SIZE + header.dataSize,
-  };
-}
-
-/**
- * Dump a GRUP to a debug file
- */
-export function dumpGrupToFile(
-  buffer: Buffer,
-  offset: number,
-  grupSize: number,
-  grupLabel: string
-) {
-  const header = `\n=== GRUP DUMP START ===\n`;
-  const footer = `\n=== GRUP DUMP END ===\n\n`;
-
-  const dump = [
-    header,
-    `Label: ${grupLabel}`,
-    `Offset: ${offset}`,
-    `Size: ${grupSize}`,
-    `Hex Dump:`,
-    hexDump(buffer, offset, grupSize + RECORD_HEADER.TOTAL_SIZE, 16).join("\n"),
-    footer,
-  ].join("\n");
-
-  fs.appendFileSync(GRUP_DEBUG_FILE, dump);
-}
